requirements.txt
aiohttp>=3.9.0
streamlit>=1.28.0
python-dotenv>=1.0.0
streamlit-autorefresh>=0.2.0
py-clob-client>=0.34.0
